* Make GET_INT_ARG() support 64 bit ints, for future use * Make $fseek() support 64 bit file offsets (yay!) * Begin converting some time_t uses over to 64 bits, Just In Case * Make $numsort(), $strtol(), $tobase() support 64 bit ints. * Make $stat() support 64 bit ints where possible.

+++ include/files.h
@@ -22,8 +22,8 @@
 	int	file_valid		(int);
 	int	file_rewind		(int);
 	int	file_error		(int);
-	int	file_seek		(int, long, const char *);
-	int	file_tell		(int);
+	int	file_seek		(int, off_t, const char *);
+	intmax_t file_tell		(int);
 	int	file_skip		(int, int);
 
 	char *	dbmctl			(char *);
